Add testcase to verify non expansion of '?list: item+' rules if more than one item matches
When using the EXPAND1 (?) operator on a rule you don't expect it to
expand if it contains more than a single item, regardless of how it came
to contain more than a single item. I.e. if the ?list rule uses a EXPAND
(@) @item rule: the list's items (children) shouldn't expand after the
list had already been expanded.

I.e. in the case of >= 2 matches these rules are expected to be
equivalent:

?list: item+;
?list: item | item item+ ;

This testcase checks for that, but currently that is *not* the case.

Signed-off-by: Giel van Schijndel <me@mortis.eu>