Flatten recursive #rules in-place to significantly reduce tree-depth
Same solution as 8d4860b5bd (that was for in-place expansion of @rules
and ?rules).

In short: without this change recursive #rules marked for FLATTENing can
cause ASTs with unbounded depth. Due to a recursive `repr()`
implementation this can cause stack overflows when running with
`debug=True`. This commit prevents this problem.

Signed-off-by: Giel van Schijndel <me@mortis.eu>