refactor parser-rule actions to use same definition for all rules
I.e. use the same action when matching a parser-rule with flags (?#@) as
without. Due to previous changes the rule for EXPAND/EXPAND1/FLATTEN was
already suitable to be used for the non-flag case. As a result the only
changes are indentation and comments along with the removal of the
non-flag-only action.

Signed-off-by: Giel van Schijndel <me@mortis.eu>