Stay closer to regular pickle protocol
Pickle by default pickles obj.__dict__, __getstate__ overrules that.
Rather than returning something completely incompatible with __dict__,
return a shallow copy of __dict__ with all contained weak references
removed. This gets rid of the 'parent' attribute in a way compatible to
the default __setstate__ implementation and as such removes the need to
provide a non-default implementation for that.

Additionally move this pickle-protocol behaviour to a mixin-class to
allow easy reuse where necessary.

Signed-off-by: Giel van Schijndel <me@mortis.eu>